Make the status line look like with Airline.
Manually build a status line that looks like the
one I had configured with Airline.

For now, only the colors for Solarized Dark are
defined.

More or less, the code to manually build the status
line is no longer than the code needed to configure
Airline. And I have complete control on how the status line is built!