Fix bug: single window with inactive status line.
The bug happened when a window was split (ctrl-w s), then a window except the top (window nr 1)
was made active, and then all the other windows
were closed (ctrl-w o).

Putting the logic for checking if a window is active out of the status line building function
and into RefreshStatus() fixes the problem.