Fix problems with Ledger entry command.
The code for LedgerEntry() is derived from

https://github.com/oppaus/vim-ledger/commit/512bb6f22b2fa769a849ad21ca8fbc9802ab1b78

Now the text written by the user is deleted and
replaced by Ledger’s entry. The whole line is
considered rather than the word under the cursor
(so the cursor does not need to be over a word),
and the ctrl-t mapping works in Insert mode, too.

This seems to work very well!