Use ,1 and similar to switch between tabs.
I’ve seen that using ,1 to go to buffer 1 (and
similarly for the other mappings) is not particularly
useful, since after opening several files, buffer
numbers quickly increase above 10. So, I’ve decided
to use the default :1b and similar to switch between
buffers, and use ,1 and similar mappings to switch
between tabs.