Add fo-=o to vimrc’s modeline.
Remove ‘o’ from formatoptions so that Vim does not
automatically start a comment after opening a new
line from a comment line.

Note that this is only for vimrc.
We might consider setting the option in vimrc in the
future (but keep in mind that plugins may overwrite
the setting in vimrc, as is the case for files of
‘vim’ type, whose formatoptions are re-set by vim.vim).

See also http://stackoverflow.com/questions/16030639/vim-formatoptions-or/23326474#23326474