Do not escape command parts, unless expansions.
Leave to the caller the responsibility of correctly
escaping a command. Why is this better?

Consider a situation in which you want to run a
command that takes a string with spaces as an option,
e.g.:

    foo -x ‘abc def’

If the parts were escaped by RunShellCommand(),
the above would become:

    ‘foo’ ‘-x’ ‘’’abc’ ‘def’’’

which is wrong.