Use H, L and M to switch buffers.
As a rule of thumb, I want all custom mappings to
start with <Leader> or <C->, with a preference for
the former. So far, I have respected such rule
with a few of exceptions (notably, Undotree and Tagbar
are activated by function keys).

But using H and L to go back and forth between buffers is very convenient and fast, especially compared to
,p (ok, I could use ,m to make typing easier…).
Besides, the default function of H, L and M is not
useful in my configuration, in which the cursor
stays vertically centered. So, I think that it is
a good idea to make a further exception to my rule
and remap them to something more useful.