Rethink the implementation of the status line.
I’m trying to keep it simple: no custom window
variables, no autocmd, no local status lines (so
plugins, e.g. CtrlP, have an opportunity to define
their own status lines).

The (global) status line is defined as follows:

    %!BuildStatusLine(winnr())

and never changed (unless the status line is disabled).
BuildStatusLine() always gets the number of
the currently active window. The function builds
a status line that invokes a series of functions
using %{f()} syntax. Such functions know which window
is active because they may get such information from
BuildStatusLine(). Besides, since they are invoked
in %{} context, they know for which window they
are building the status line (see `:h statusline`).

If you do not need to get fancy, this works just fine
to create different status lines for the active window
and the inactive ones. But I want to get a bit fancy:
I’d like to use different colors for different parts
and  highlight the mode I am in, as before. This
turns out to be a bit tricky, because the return
value of a function called in %{} context is taken
literally (so, such functions cannot return strings
that change the highlight group).

Alternative status lines are not re-implemented yet.