Adjust timeout for search().
Searching for trailing spaces is significantly
slower than than other two searches, so we give it
a bit more time.

Note, however, that in my tests a trailing space
at the end of a ~13MB file takes more than 2s to be
found (so, with the current settings it would
not be detected).