Customize the tab line.
Show the tab number, to make it easy to switch
between tabs with my custom mappings (\1, \2, …).

Show only the file name, not the (shortened) path.

Make tab labels a bit wider.

Use a custom modified symbol, as in the status line.

Try to keep everything as efficient as possible by
avoiding commands (:for, :if, …) and using functions
instead.