Fix bugs in tab autocompletion for Ledger files.
Pressing tab twice would cause the expansion of
the <tab> mapping to be printed literally instead of
being executed. Now that is fixed.

Instead of using `<c-o>:` we now use `<c-r>=`. The
only tricky thing is that we need to move the cursor
one character to the left before invoking
ledger#align_amount_at_cursor() (<c-o> used to do it
for us). Correspondingly, to leave the cursor at
the end of the line, we need to return `<c-o>A`
(temporarily switch to normal mode to execute A)
after the amount has been aligned. This seems to
work perfectly.

Btw, for some reason <S-Tab> also works in the
pop-up menu, even if we have not mapped it.