Typeset LaTeX files asynchronously (clientserver).
Spawning a process in the background and having a
callback function called when the process is over
is fairly easy if Vim is built with clientserver.
The basic structure of the command is:

!(<command>; mvim —remote-send ‘<callback>’) &

To make this work it seems necessary to redirect both
the stdout and stderr of the command to /dev/null.
Other than that, it works flawlessly.

It is possible to pass the exit status of the command
back to Vim through the $? shell variable. So, the
callback function may determine the result of the
command and act accordingly.