Use a global variable to store the cached mode.
The current mode is global, not window specific,
so it is better to store it in a global variable.

Note that the “w:mode” attribute is only needed
to be able to call s:updateSepMode() without an
explicit :call command: it is never used elsewhere.
In fact, we might get rid of it, and use w:active’s
value to do the same computation…