Optimize s:getBackground().
This version is significantly faster than the previous one (it might be nearly as twice as fast).

It seems that calling synIDattr() and co. is quite
expensive, so it is better to reduce the number of such calls.

Besides, according to my benchmarks, it is not worth to pursue further optimization (e.g., by inlining getBackground()).