Remove the lf_mode window attribute.
This is not used anywhere, and we can slightly
optimize the computation of the status line by
calling updateHighlightGroups() (now renamed to
updateStatusLineHightlight()) when computing the
lf_active attribute.

Also, note that updateStatusLineHighlight() does not
cache the current mode any longer: the value is
cached in SetupStl() before calling updateStatusLineHighlight().