Enable lazy loading of plugins.
Pathogen has a g:pathogen_blacklist variable that may be used to specify
a list of plugins that should not be loaded at startup (btw, it also
supports an undocumented $VIMBLACKLIST variable to temporarily disable
plugins). Currently, I use that list to disable YouCompleteMe, which
significantly increases the startup time, and which I need only when
using some programming languages (e.g., Python).

Enabling plugins at runtime is “super-tricky” in general:

    https://github.com/tpope/vim-pathogen/issues/78

(YCM being a good example, since it requires to call a function to
enable it), but I don’t need a completely general approach in my
personal vimrc, so s:loadPlugin() should be good enough.