Use quickfix list for Ledger errors and reports.
This is a major revision of my Ledger functions. Now, it is possible to
use the quickfix list to display register reports, so that one can
easily navigate through postings. This is useful in particular for
reconciling.

All Ledger functions now are completely self-contained (they do not
depend on my Shell command in .vimrc).

The idea is to move this code into the Ledger plugin and submit a pull
request.