Enabling resizing splits with the mouse inside tmux.
Quoting from

    http://superuser.com/questions/549930/cant-resize-vim-splits-inside-tmux

"It appears that dragging the status line to resize a split is not
possible when the Vim option `ttymouse` is `xterm`; it does work when the
value is `xterm2` though. The latter value configures Vim to ask for an
extended mouse reporting mode that (among other things) provides better
dragging support. This extended mode only works with newer versions of
xterm (and other compatible terminal emulators, including tmux), so it
is not the default value."

Setting ttymouse is not needed in NeoVim because the option has been
removed (resizing with the mouse works out of the box in NeoVim).

If Vim is run from a terminal (outside tmux), ttymouse is already set
to `xterm2`, at least in all the environments I have tried.