Further optimize the status line.
Remove lf_bufnr: I don't know why I hadn't realized before that it is
not needed.

Show the file name next to the mode, and print the directory containing
the file after the modified/read only symbols. This looks cleaner to me,
especially because the position of the file name does not depend on the
length of the path.