Refactor vimrc, use autoload.
This is mostly a clean up and refactoring of vimrc. All functions that
are not needed immediately after startup have been put into autoload
files (not that this improves performance in any measurable way, but the
vimrc is a bit tidier).

Along the way I have removed unused code, changed the layout of some
other parts, reordered the code, and made some (minor) functional
changes.