Add Sneak plugin.
This seems to me a much better implementation than EasyMotion (and with
less eye-candy). It works with different language maps, it does *not*
modify the buffer, highlights targets significantly faster, it
integrates well with Surround and Repeat, etc etc. Besides, it loads in
1/4 the time (~1.2ms): still not the fastest loading plugin in the
world, but fine.

Really, I don't know why this isn't more popular (currently, 591 stars
vs 2260 stars for EasyMotion, on Github).