Put some LaTeX functions into autoload.
This will allow me to share such functions with other TeX engines
(namely, ConTeXt).

I have also changed the clean function so that it does not use latexmk.
This is so that I have total control over which files are deleted, and I
may use the same functions both for LaTeX and ConTeXt. The list of files
to be deleted is preliminary and should be refined.