Update LuaLaTeX callback function and fix bugs.
The callback function works fine in Vim. Apparently, there is a bug in
NeoVim, because if the function is not already loaded when jobstart() is
called, NeoVim does not find it. As a workaround, I call the function
explictly, passing a dummy ‘load’ event, which causes the function to
(be loaded and) return immediately.