Moves themes into pack/themes/opt.
Start using Vim’s package functionality by moving all themes under
`pack/themes/opt`. Preserve compatibility with older Vim’s versions by
updating Pathogen’s paths accordingly.

Note that `:colorscheme foo` searches `colors/foo.vim` in `runtimepath`
first, then in `pack/*/start/*` and finally in `pack/*/opt/*` (even when
the colorscheme is loaded in `vimrc`, that is, before plugins are
actually loaded).
This means that a colorscheme will be found even if it is not in
`runtimepath`. So, it is not strictly necessary to clobber `runtimepath`
with paths to colorschemes. That’s why we put themes under `opt` rather
than `start`. There are a couple of caveats, though:

1. autoload functions in a theme folder will not be found unless the
   theme folder is added to `runtimepath` with `:packadd!`. This is not
   a big deal for me, because usually the autoload folder provides
   functions for Airline or Lightline, which I do not use.
2. `~/.vim/after` must not contain a `colors/mytheme.vim` file,
   otherwise that file will be sourced instead of the theme (because
   `runtimepath` is searched first and `runtimepath` always contains
   `~/.vim/after`). This is why I have renamed `~/.vim/after/colors`
   into `~/.vim/after/themes`.

Note that putting a colorscheme inside `pack/themes/start` would not
solve (2), because pack paths are processed *after* `vimrc` has been
sourced, that is, after a `:colorscheme` command in `vimrc` is executed
(it would solve (1), of course). To solve (2), we should put this in
`vimrc`:

    packadd! mytheme
    colorscheme mytheme

The `packadd!` command would prepend mytheme’s path to `runtimepath`, so
that the subsequent `colorscheme` command would certainly find it. I
prefer to leave colorschemes inside `opt` and make sure that they are
not found in `runtimepath`.