Avoid clobbering hlsearch with tp in TeX buffers.
The way the tp operator was implemented (with ?) was causing the search
pattern to be updated. If hlsearch was on, the search pattern was
highlighted. Not nice.

Now we use search(), which does not affect the search pattern, both for
LaTeX and ConTeXt. This commit also removes the duplicated regex.

As a bonus, now a paragraph is matched even when the cursor is on the
line before the paragraph. For example, given

\begin{foo}
xxx
\end{foo}

gqtp or vtp will act on xxx even when the cursor is on \begin{foo}. This
is especially useful for list items, which I lay out as follows:

\begin{itemize}
  \item
  Item One
  \item
  Item Two
\end{itemize}

I can select the item’s text even when the cursor is on the preceding
\item line.