Define an autoload function for :Ctags.
Use better defaults and provide some feedback. The canonical way to use
this command is simply to type :Ctags (with no arguments), to create the
tags file for the directory of the current buffer (would it be better to
use the current working directory?). Optional arguments are passed to
ctags: they may be used, for example, to exclude further directories.