Make TeX callback more robust.
Now, I can freely switch to other buffers while a TeX job is running
(before this commit, that would have prevented Vim to locate the log
file).

I have removed the code to restore the local working directory, so now
the callback has a side effect. This should not have adverse effects in
practice, and the benefit is (besides simplified code) that the file
paths in the quickfix list are always relative to the directory of the
TeX document.