Be smarter creating tags by using tagfiles().
With this commit, I can safely type :Ctags and be sure that the tags
file will be created at the right place, no matter what the current
working directory is and even if the current file is not located at the
project's root.

Before this commit, lf_text#ctags() would modify the local working
directory (bad side effect). Now the function uses tagfiles() to detect
plausible locations for creating tags. It works as follows:

- if tagfiles() reports no path, then ask the user where a new tags file
  should be created. Candidates are the current working directory, the
  directory containing the current file, or something else;
- if tagfiles() contains exactly one path, then use that and do not
  bother the user with questions; in my workflow, this is by far the
  most common case;
- if tagfiles() contains more than one path, let the user choose one
  from a list. I could cache the user's choice, but, at least for now,
  no caching is performed, because this case shouldn't happen often.