Get info on the highlight group at current position.
Implement a function to identify the highlight group at the current
cursor position. Add another function that adds/removes an autocmd to
automatically print such information as the cursor moves.

The information is printed under the status line and looks like this:

    rubyClass ⊂ rubyBlock xxx rubyClass → rubyDefine    fg=10/#586e75    bg=15/#fdf6e3

From left to right you may find:

1. the synstack (highlight group containment relationships);
2. a sample of the identified highlight group (xxx);
3. the name of the highlight group; if the highlight group links to
   other highlight groups, then the "base" highlight group is also
   shown (`rubyDefine` in the example above);
4. the cterm/gui values of the foreground color and the background
   color, respectively.

Transparent groups are supported, too.

I have added a try-catch block because I have found a circumstance in
which the value returned by synIDattr() is not correct. More
specifically, if CtrlP is opened while the autocmd of this commit is
active, a "cannot allocate color" error is raised. This is because, for
some reason, synIDattr(synid, "fg", "gui") returns the same value as
synIDattr(synid, "fg", "cterm"), i.e., an integer rather than a #RGBRGB
color.