Fix lf_git#diff().
The new function I am using to send the output of a command to a new
buffer (lf_job#to_buffer()) does not switch the focus to the newly
created window, as the previous function did. So, I need to reorder the
diffing commands.

I have also added `silent` to `file` to suppress a Vim prompt.