Revert e17927d3.
The reason I'm removing these is that, after some use, I have found the
following too annoying: even when there is no match for the word under the
cursor, the text scrolls vertically to center the line where the cursor
is. This gives me the wrong impression that I have jumped to a match.