Update functions for asynchronous TeX typesetting.
I have added functions to check the status of TeX jobs and to kill all
running TeX jobs. The callback is more robust (it works even if the user
moves to a different window/tab or closes the buffer while the
typesetting job is running).

These work in both Vim and NeoVim.