Add support for <c-n> and <c-p> completions.
This subsumes the incl method, because I make sure in my vimrc that the
'complete' option contains 'i' (which is the default in Vim, but not in
NeoVim).

I could also add ']' to 'complete', but I prefer to use tags only if the
other completion methods fail.

So, the current completion chain is as follows:

1) first, try omnicompletion;
2) try local keyword completion (keywords int the current buffer);
3) try 'complete' completion (this and other buffers, included files,
   spell dictionary);
4) search tags, if available;
5) search words from 'dictionary', if available.

I could probably add 'k' to 'complete' and get rid of 5). I need to
experiment a bit.