Add draft of ConTeXt and SQL abbreviations.
I plan to remove UltiSnips, since with Vim abbreviations and a few
tricks I think I can cover all my needs. The tricks include setting
marks to be able to jump between different parts of a snippet (already
done in the `create function` SQL snippet) and adding jumps to the jump
list to be able to move back and forth with <c-o> and <c-i> (to be
done).