Use pstopdf in :hardcopy.
For some reason, Preview.app does not convert PostScript files any more.
If I attempt to open a PostScript file, I get the following message:

    PostScript Conversion Error

    The PostScript file “<filename>” could not be converted to
    a PDF file.

I do not print files very often, but it used to work some time ago (and
some OS releases before). In macOS Sierra I am getting this error.

To work around the problem, I now convert the PostScript file generated
by Vim with `pstopdf`, and open the resulting PDF with Preview.

This commit also fixes a minor issue whereby printed files were not
deleted. Note that it is not a problem for Preview if the PDF file is
deleted after it has been opened (at least if you just want to print
it).