Remove `set lazyredraw` to improve cursor appearance.
Without lazyredraw on, it seems that the "flickering cursor" problem
mentioned in this µcomplete's commit:

https://github.com/lifepillar/vim-mucomplete/commit/8d2845baa187ac50b0da878b82e322d38e1859f4

disappears. For the record, I have set the terminal cursor to a bar when
in Insert mode and to a block in Normal mode. During typing, however,
the cursor may intermittently become blockwise if µcomplete's
autocompletion is on. In the commit mentioned above, I had mistakenly
attributed the misbehaviour to feedkeys(), but after some testing,
I have come to the conclusion that lazyredraw is the culprit.

With lazyredraw off, repeating macros may become noticeably slower, so
maybe I will define a toggle for lazyredraw.