Configure µcomplete for Ledger.
Here, <tab> aligns a commodity if it is pressed after an amount,
otherwise it invokes µcomplete as usual.

Omni-completion is set up so that it is used either when two alphabetic
characters or colons are in front of the cursor, or at the start of
a line (to insert the current date). Keyword completions are restricted
to alphabetic characters to avoid auto-completing digits.