Improve <a-...> mappings.
There was a problem with the alt mappings before this commit: typing
<esc> quickly followed (i.e., within a second) by one of the mapped keys
(e.g., <esc> followed by j), would trigger the mapping. Quite annoying.
Also, typing <esc> in Command mode would cause Vim to wait for a second
for another key. This was a minor annoyance, but I have fixed that, too.

Now, I have discovered a trick to distinguish <esc> (relatively quickly)
followed by a key, and an alt mapping. See:

    https://stackoverflow.com/questions/6778961/alt-key-shortcuts-not-working-on-gnome-terminal-with-vim/10216459

I have implemented the idea with alt-arrows in Command mode (to jump
between words), and with alt-hjkl in Normal mode to move between
windows. I have tested it and it works inside and outside tmux, no
matter what the terminal settings for the Option key are (i.e., in
Terminal.app, “Use Option as Meta key” may be checked or unchecked; in
iTerm2, alt may be interpreted as Normal or Esc+).

In NeoVim things work completely differently (see :help :map-alt-keys).
Only the mappings for alt-arrows seem to be necessary (and they are
different from Vim's mappings).

Finally, everything works provided that ttimeoutlen is a sufficiently
low value. For now, I have set it to 50, but I'll probably lower it even
more.