Implement functions to comment/uncomment code.
This more or less replaces the functionality of the Commentary plugin
(at least, the part of it I was using). In some respects, it works even
better (with all due respect to T. Pope's outstanding work). For
example, when commenting lines with decreasing indentation, Commentary
does not align the comments, so this:

        xxx
      xxx
    xxx

becomes

        " xxx
      " xxx
    " xxx

instead of

    "    xxx
    "  xxx
    " xxx

This commit produces the latter result (possibly with some loss of
efficiency compared to Commentary—I haven't performed any benchmark).
There are most likely several bugs in this commit, but with time I will
discover them and iron them out.

Use Q to comment and gQ to uncomment. Both mappings are operators, so
they can be used in Visual mode or followed by a motion.

To comment the current line, you may use any motion that stays in the
line. I find QW the fastest to type. Uncommenting the current line is
done in a similar way.

I have discovered that the trickiest part of this commenting stuff is
determining the comment delimiters. Some plugins hard-code the comment
strings (CommentAnyWay). Commentary and other plugins (e.g.,
NERDcommenter), use 'commentstring'. So do I, for now: it is simple
enough and works reasonably well. In the future, I might implement more
sophisticated functions, e.g., using 'comments' to find the possible
delimiters (say, for uncommenting better). It would be nice to have
block commenting/uncommenting, too.