Make sending keys to the terminal more robust.
See this discussion:

    https://github.com/vim/vim/issues/1985

The current solution seems to work well both in terminal Vim (as of
v8.0.946) and in MacVim (8.0.972).

Using term_wait() or WaitFor() does not seem necessary in MacVim (in
fact, call term_wait() has the side effect of making the shell exit, see
https://github.com/macvim-dev/macvim/issues/542), but it is required in
terminal Vim (otherwise some characters may be lost or the text is
messed up). Also note that calling WaitFor() after the for loop is not
enough.