Update terminal mappings for <s-left> and <s-right>.
A drawback of the previous mappings is that, in an instance of Vim run
inside a terminal window (e.g., when executing a `git commit`), I needed
to type Esc twice to return to Normal mode. The new mappings avoid that
issue, and they work mostly the same otherwise (one difference is that
Shift-arrows work the same as Alt-arrows in a Vim terminal window, but
that's ok with me).

This means that, together with these other environmental settings in my
vimrc:

    set <s-left>=^[b  " There's a literal Esc (^[) here and below
    set <s-right>=^[f

and with the following settings in the terminals I use (Terminal.app and
iTerm2):

    ⇧ ← mapped to \033[1;2D
    ⇧ → mapped to \033[1;2C
    ⌥ ← mapped to \033b
    ⌥ → mapped to \033f

I can now use Alt-arrows to jump between words in Vim in all modes, and
in Vim terminal windows as well. This works even in instances of Vim run
inside a terminal window run inside a Vim instance run inside etc…
Besides, now Esc returns to Normal mode immediately under all
circumstances.