Restore window layout after closing interactive finder.
Use winrestcmd() to save the commands needed to restore the current
window layout before opening the interactive filter's window. This is
necessary because in some cases (in particular, when a window with
winfixheight set exists) the height of some windows may be changed
unexpectedly.

1. Consider the following layout:

:-----------------:
:                 :
:   some window   :
:                 :
:-----------------:
: quickfix window :
:-----------------:

Before this commit, when opening the finder (below the quickfix window),
the finder's window would be taller than 10 lines (see also
https://github.com/vim/vim/issues/2999). This is addressed by replacing

    botright 10new

with

    botright 1new | 9wincmd +

2. Consider this layout:

:--------:--------:
:        |        :
:        |        :
:        |        :
:-----------------:
: quickfix window :
:-----------------:

Before this commit, after opening and closing the finder (below the
quickfix window), the quickfix window would not retain its height: it
would become taller (see also https://github.com/vim/vim/issues/2998).
This is addressed by this commit by saving (and then restoring) the
window layout with winrestcmd().