Update the status line to use new Vim features.
Use g:statusline_winid, if available, to detect
whether we are drawing in the active status line.

Move the code for the old status line into a
separate file.

This new definition seems somewhat faster: a quick
test shows that a status line is drawn in ~25µs vs
~35µs for the old status line.

It is also more flexible, because active and
inactive status lines may be defined separately.
For instance, in this new definition, the encoding
is not shown in inactive status lines.