Modify the way 'mixed indent' is determined.
Use a stricter definition than before: a mixed indent warning should be
given only when the same document has a line starting with at least two
spaces and another line starting with a tab. Tabs and spaces in the same
line are tolerated: this is because, when using hard tabs, tabs are used
for indentation and spaces for alignment, so they may well be together
in the same line. 

Why search for two spaces, not one? Again, when using hard tabs, one may
encounter situations like the following C snippet:

    /*
     * A comment.
     */
    int main() {
		return 0;
    }

So, to align the asterisks one must prepend a single space. Since such
space is used for alignment and not indentation, it should not produce
warnings.

The current method does not raise any warning if a line starts with
spaces following by tabs. This should probably be fixed.