Use TeXShop as the default TeX previewer.
Starting with version 4.25, it is possible to use TeXShop for both
forward and backward search. Since, for some reason, Skim's synctex
functionality has stopped working for me recently, I am switching to
TeXShop (which also seems faster).

For this to work, you need to put two shell scripts in /usr/local/bin:

- `othereditor` (this name cannot be changed, it is required by TeXShop)
  with the following content:

        #!/bin/sh
        /usr/local/bin/mvim --remote-silent +"$1" "$2"

- `texshop.sh` (the name is arbitrary) which embeds a small AppleScript
  script to open TeXShop and go to the requested position in a document.
  For now, I am using the script that can be found in TeXShop.app >
  Help > Changes, and which I report here verbatim for the sake of
  completeness:

        #!/bin/bash
        MyShellVar=$1
        MyShellVas=$2
        MyShellVat=$3
        osascript <<EOD
          tell application "TeXShop"
            activate
            set the front_document to the front document
            set MyAppVar to $MyShellVar
            set MyAppVas to $MyShellVas
            set MyAppVat to "$MyShellVat"
            tell front_document
              sync_preview_line theLine MyAppVar
              sync_preview_index theIndex MyAppVas
              sync_preview_name theName MyAppVat
              return 0
            end tell
          end tell
        EOD

You also need to check "Configure for External Editor" in TeXShop's
Prefereces and set a hidden preference:

    defaults write TeXShop OtherEditorSync YES

Works like a charm, at least with LaTeX (with ConTeXt I haven't tried
yet). Note that the PDF document must already be opened in TeXShop for
searches to jump into it.

Of course, the document must be typeset with SyncTeX enabled!