Escape :grep's argument.
This still requires escaping special characters on the command line.
For instance, :Grep #foo does not work (because # is expanded before invoking the function),
but :Grep \#foo is fine. The \ is stripped before passing '#foo' to the function; inside the function,
fnameescape() escapes # again for :grep!, then shellescape() escapes the input for the shell,
(so that the shell see '#foo', with quotes).